+++ core/market/src/db/model.rs
@@ -7,7 +7,7 @@ mod proposal;
 mod proposal_id;
 mod subscription_id;
 
-pub use agreement::{Agreement, AgreementId, AgreementState, AppSessionId};
+pub use agreement::{check_transition, Agreement, AgreementId, AgreementState, AppSessionId};
 pub use agreement_events::{AgreementEvent, AgreementEventType, NewAgreementEvent};
 pub use demand::Demand;
 pub use negotiation_events::{EventError, EventType, MarketEvent};
