+++ core/market/src/db/model/agreement.rs
@@ -10,6 +10,7 @@ use ya_client::model::market::offer::Offer as ClientOffer;
 use ya_client::model::{ErrorMessage, NodeId};
 use ya_diesel_utils::DbTextField;
 
+use crate::db::dao::StateError;
 use crate::db::model::{OwnerType, Proposal, ProposalId, SubscriptionId};
 use crate::db::schema::market_agreement;
 
@@ -186,4 +187,33 @@ impl From<AgreementState> for ClientAgreementState {
             AgreementState::Terminated => ClientAgreementState::Terminated,
         }
     }
+}
+
+pub fn check_transition(from: AgreementState, to: AgreementState) -> Result<(), StateError> {
+    log::trace!("Checking Agreement state transition: {} => {}", from, to);
+    match from {
+        AgreementState::Proposal => match to {
+            AgreementState::Pending => return Ok(()),
+            AgreementState::Cancelled => return Ok(()),
+            AgreementState::Expired => return Ok(()),
+            _ => (),
+        },
+        AgreementState::Pending => match to {
+            AgreementState::Cancelled => return Ok(()),
+            AgreementState::Rejected => return Ok(()),
+            AgreementState::Approved => return Ok(()),
+            AgreementState::Expired => return Ok(()),
+            _ => (),
+        },
+        AgreementState::Cancelled => (),
+        AgreementState::Rejected => (),
+        AgreementState::Approved => match to {
+            AgreementState::Terminated => return Ok(()),
+            _ => (),
+        },
+        AgreementState::Expired => (),
+        AgreementState::Terminated => (),
+    };
+
+    Err(StateError::InvalidTransition { from, to })
 }
